Use ServiceLoader to load the appropriate RemoteTestNG
