+++ common/src/main/java/org/testng/remote/AbstractRemoteTestNG.java
@@ -1,18 +1,14 @@
 package org.testng.remote;
 
-import com.beust.jcommander.JCommander;
-import com.beust.jcommander.ParameterException;
 import org.testng.CommandLineArgs;
 import org.testng.ISuite;
 import org.testng.ISuiteListener;
 import org.testng.ITestRunnerFactory;
 import org.testng.TestNG;
-import org.testng.TestNGException;
 import org.testng.collections.Lists;
 import org.testng.remote.strprotocol.*;
 import org.testng.xml.XmlSuite;
 
-import java.util.Arrays;
 import java.util.List;
 
 import static org.testng.internal.Utils.defaultIfStringEmpty;
@@ -22,17 +18,8 @@ import static org.testng.internal.Utils.defaultIfStringEmpty;
  *
  * @author Cedric Beust <cedric@beust.com>
  */
-public abstract class AbstractRemoteTestNG extends TestNG {
+public abstract class AbstractRemoteTestNG extends TestNG implements IRemoteTestNG {
   private static final String LOCALHOST = "localhost";
-
-  // The following constants are referenced by the Eclipse plug-in, make sure you
-  // modify the plug-in as well if you change any of them.
-  public static final String DEBUG_PORT = "12345";
-  public static final String DEBUG_SUITE_FILE = "testng-customsuite.xml";
-  public static final String DEBUG_SUITE_DIRECTORY = System.getProperty("java.io.tmpdir");
-  public static final String PROPERTY_DEBUG = "testng.eclipse.debug";
-  public static final String PROPERTY_VERBOSE = "testng.eclipse.verbose";
-  // End of Eclipse constants.
 
   protected ITestRunnerFactory m_customTestRunnerFactory;
   private String m_host;
@@ -41,19 +28,59 @@ public abstract class AbstractRemoteTestNG extends TestNG {
   private Integer m_port = null;
 
   /** Port used for the serialized protocol */
-  private static Integer m_serPort = null;
+  private Integer m_serPort = null;
 
   /** Protocol used for inter-communication */
-  private static String m_protocol;
+  private String m_protocol;
 
-  private static boolean m_debug;
+  private boolean m_debug;
 
-  private static boolean m_dontExit;
+  private boolean m_dontExit;
 
-  private static boolean m_ack;
+  private boolean m_ack;
 
+  @Override
+  public void dontExit(boolean dontExit) {
+    m_dontExit = dontExit;
+  }
+
+  @Override
+  public void setDebug(boolean debug) {
+    m_debug = debug;
+  }
+
+  @Override
+  public void setAck(boolean ack) {
+    m_ack = ack;
+  }
+
+  @Override
   public void setHost(String host) {
     m_host = defaultIfStringEmpty(host, LOCALHOST);
+  }
+
+  @Override
+  public void setSerPort(Integer serPort) {
+    m_serPort = serPort;
+  }
+
+  @Override
+  public void setProtocol(String protocol) {
+    m_protocol = protocol;
+  }
+
+  @Override
+  public void setPort(Integer port) {
+    m_port = port;
+  }
+
+  @Override
+  public void configure(CommandLineArgs cla) {
+    super.configure(cla);
+  }
+
+  public static void validateCommandLineParameters(CommandLineArgs args) {
+    TestNG.validateCommandLineParameters(args);
   }
 
   private void calculateAllSuites(List<XmlSuite> suites, List<XmlSuite> outSuites) {
@@ -67,7 +94,7 @@ public abstract class AbstractRemoteTestNG extends TestNG {
   public void run() {
     IMessageSender sender = getMessageSender();
     final MessageHub msh = new MessageHub(sender);
-    msh.setDebug(isDebug());
+    msh.setDebug(m_debug);
     try {
       msh.connect();
       // We couldn't do this until now in debug mode since the .xml file didn't exist yet.
@@ -141,80 +168,6 @@ public abstract class AbstractRemoteTestNG extends TestNG {
    * automatically.
    */
   protected abstract ITestRunnerFactory buildTestRunnerFactory();
-
-  protected static void main(AbstractRemoteTestNG remoteTestNg, String[] args) throws ParameterException {
-    CommandLineArgs cla = new CommandLineArgs();
-    RemoteArgs ra = new RemoteArgs();
-    new JCommander(Arrays.asList(cla, ra), args);
-    m_dontExit = ra.dontExit;
-    if (cla.port != null && ra.serPort != null) {
-      throw new TestNGException("Can only specify one of " + CommandLineArgs.PORT
-          + " and " + RemoteArgs.PORT);
-    }
-    m_debug = cla.debug;
-    m_ack = ra.ack;
-    if (m_debug) {
-//      while (true) {
-        initAndRun(remoteTestNg, args, cla, ra);
-//      }
-    }
-    else {
-      initAndRun(remoteTestNg, args, cla, ra);
-    }
-  }
-
-  private static void initAndRun(AbstractRemoteTestNG remoteTestNg, String[] args, CommandLineArgs cla, RemoteArgs ra) {
-    if (m_debug) {
-      // In debug mode, override the port and the XML file to a fixed location
-      cla.port = Integer.parseInt(DEBUG_PORT);
-      ra.serPort = cla.port;
-      cla.suiteFiles = Arrays.asList(new String[] {
-          DEBUG_SUITE_DIRECTORY + DEBUG_SUITE_FILE
-      });
-    }
-    remoteTestNg.configure(cla);
-    remoteTestNg.setHost(cla.host);
-    m_serPort = ra.serPort;
-    m_protocol = ra.protocol;
-    remoteTestNg.m_port = cla.port;
-    if (isVerbose()) {
-      StringBuilder sb = new StringBuilder("Invoked with ");
-      for (String s : args) {
-        sb.append(s).append(" ");
-      }
-      p(sb.toString());
-//      remoteTestNg.setVerbose(1);
-//    } else {
-//      remoteTestNg.setVerbose(0);
-    }
-    validateCommandLineParameters(cla);
-    remoteTestNg.run();
-//    if (m_debug) {
-//      // Run in a loop if in debug mode so it is possible to run several launches
-//      // without having to relauch RemoteTestNG.
-//      while (true) {
-//        remoteTestNg.run();
-//        remoteTestNg.configure(cla);
-//      }
-//    } else {
-//      remoteTestNg.run();
-//    }
-  }
-
-  private static void p(String s) {
-    if (isVerbose()) {
-      System.out.println("[RemoteTestNG] " + s);
-    }
-  }
-
-  public static boolean isVerbose() {
-    boolean result = System.getProperty(PROPERTY_VERBOSE) != null || isDebug();
-    return result;
-  }
-
-  public static boolean isDebug() {
-    return m_debug || System.getProperty(PROPERTY_DEBUG) != null;
-  }
 
   private String getHost() {
     return m_host;
