+++ pom.xml
@@ -17,6 +17,7 @@
 
     <properties>
         <gson.version>2.5</gson.version>
+        <auto-service.version>1.0-rc2</auto-service.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.7</maven.compiler.source>
         <maven.compiler.target>1.7</maven.compiler.target>
@@ -28,6 +29,12 @@
                 <groupId>com.google.code.gson</groupId>
                 <artifactId>gson</artifactId>
                 <version>${gson.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.google.auto.service</groupId>
+                <artifactId>auto-service</artifactId>
+                <version>${auto-service.version}</version>
+                <optional>true</optional>
             </dependency>
             <dependency>
                 <groupId>org.testng</groupId>
